requirements.txt
streamlit==1.31.0
tensorflow-cpu==2.15.0
librosa==0.10.1
numpy==1.24.3
Pillow==10.2.0
scikit-learn==1.4.0
soundfile==0.12.1